moved code from app.py to utils files and other clean-up
