+++ pages/create_db_entry.html
@@ -1,0 +1,9 @@
+<form action="/images" method="POST">
+
+    <label for="db_entry">Enter filename for db</label>
+    <br>
+    <input type="text" name="imageName">
+    <br>
+    <button type="submit">Submit</button>
+
+</form>
